abc-369 virtual

+++ atcoder/arc-182/b.cpp
@@ -4,36 +4,39 @@
 #define S second
 using namespace std;
 typedef unsigned long long ll;
-typedef pair<ll,ll> pll;
+typedef pair<ll, ll> pll;
 void solve()
 {
-    ll n,k;
-    cin>>n>>k;
-    ll temp=(1<<k)-1;
-    for(ll i=0; i<n; i++)
+    ll n, k;
+    cin >> n >> k;
+    ll temp = (1 << k) - 1;
+    for (int i = 0; i < n; i++)
     {
-        if(i>=(1<<(k-1)))
+        if ((ll)i >= (ll)(1 << (k - 1)))
         {
-            cout<<"1 ";
+            cout << "1 ";
             continue;
         }
-        ll ret=temp;
-        for(int j=0; j<k; j++)
+        ll ret = temp;
+        for (int j = 0; j < k; j++)
         {
-            if((i>>j)&1) ret^=(1<<(k-j-2));
+            if ((i >> j) & 1)
+                ret ^= (1 << (k - j - 2));
         }
-        cout<<ret<<' ';
+        cout << ret << ' ';
     }
-    cout<<endl;
+    cout << endl;
 }
 int main()
 {
-    ios_base::sync_with_stdio(0);cin.tie(0);cout.tie(0);
+    ios_base::sync_with_stdio(0);
+    cin.tie(0);
+    cout.tie(0);
     ll t;
-    cin>>t;
-    while(t--)
+    cin >> t;
+    while (t--)
     {
-      solve();
+        solve();
     }
     return 0;
 }
